+++ django/applications/catmaid/spatial.py
@@ -1,5 +1,8 @@
 from django.db import connection
 
+from catmaid import locks
+
+
 def enable_spatial_update_events(ignore_missing_fn:bool=False) -> bool:
     """Enable history tracking globally.
     """
@@ -13,7 +16,14 @@ def enable_spatial_update_events(ignore_missing_fn:bool=False) -> bool:
             # If the function does not exist, return silently if the missing
             # function shouldn't be reported
             return False
-    cursor.execute("SELECT enable_spatial_update_events()")
+    cursor.execute("""
+        -- Obtain an advisory lock so that this function works also in a parallel
+        -- context.
+        SELECT pg_advisory_xact_lock(%(lock_id)s::bigint);
+        SELECT enable_spatial_update_events();
+    """, {
+        'lock_id': locks.spatial_update_event_lock
+    })
     return True
 
 
@@ -30,6 +40,13 @@ def disable_spatial_update_events(ignore_missing_fn:bool=False) -> bool:
             # If the function does not exist, return silently if the missing
             # function shouldn't be reported
             return False
-    cursor.execute("SELECT disable_spatial_update_events()")
+    cursor.execute("""
+        -- Obtain an advisory lock so that this function works also in a parallel
+        -- context.
+        SELECT pg_advisory_xact_lock(%(lock_id)s::bigint);
+        SELECT disable_spatial_update_events();
+    """, {
+        'lock_id':  locks.spatial_update_event_lock
+    })
     return True